AIGC_Club_Planner/requirements.txt
fastapi
uvicorn
pydantic
langchain
python-dotenv
requests
openai
